Move deadVarElim.cs classes into separate files
